requirements.txt
pynacl
msgpack
flask
base58
pygments
qrcode
pillow
